+++ CHANGELOG.md
@@ -11,6 +11,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - `Nucleotides` and `AminoAcids` (and view types) can now be converted into `Vec`, slices, and `Cow` using `From` 
 - `NucleotidesView`, `AminoAcidsView`, and `QualityScoresView` now have inherent `slice` methods offering a longer lifetime of the returned view
 - Added `as_map` accessor to convert `ByteIndexMap` to `ByteMap`
+- Similar to the `OrFail` trait for `Result`, a `Fail` trait has been introduced to handle errors directly
 
 ### Changed